fix bug with influx. add integration tests
