+++ core/src/test/kotlin/com/malinskiy/marathon/analytics/metrics/remote/influx/KInfluxDBContainer.kt
@@ -1,0 +1,5 @@
+package com.malinskiy.marathon.analytics.metrics.remote.influx
+
+import org.testcontainers.containers.InfluxDBContainer
+
+class KInfluxDBContainer : InfluxDBContainer<KInfluxDBContainer>()
